+++ ch05/rtda/local_vars_test.go
@@ -16,8 +16,8 @@ func TestInt(t *testing.T) {
 	localVars := newLocalVars(10)
 	for i := range localVars {
 		v := rand.Int31()
-		localVars.SetInt(uint(i), v)
-		assert.Equal(t, v, localVars.GetInt(uint(i)), "")
+		localVars.SetInt(uint8(i), v)
+		assert.Equal(t, v, localVars.GetInt(uint8(i)), "")
 	}
 }
 
@@ -25,8 +25,8 @@ func TestFloat(t *testing.T) {
 	localVars := newLocalVars(10)
 	for i := range localVars {
 		v := rand.Float32()
-		localVars.SetFloat(uint(i), v)
-		assert.Equal(t, v, localVars.GetFloat(uint(i)), "")
+		localVars.SetFloat(uint8(i), v)
+		assert.Equal(t, v, localVars.GetFloat(uint8(i)), "")
 	}
 }
 
@@ -34,8 +34,8 @@ func TestLong(t *testing.T) {
 	localVars := newLocalVars(10)
 	for i := 0; i < 5; i++ {
 		v := rand.Int63()
-		localVars.SetLong(uint(i*2), v)
-		assert.Equal(t, v, localVars.GetLong(uint(i*2)), "")
+		localVars.SetLong(uint8(i*2), v)
+		assert.Equal(t, v, localVars.GetLong(uint8(i*2)), "")
 	}
 }
 
@@ -43,7 +43,7 @@ func TestDouble(t *testing.T) {
 	localVars := newLocalVars(10)
 	for i := 0; i < 5; i++ {
 		v := rand.Float64()
-		localVars.SetDouble(uint(i*2), v)
-		assert.Equal(t, v, localVars.GetDouble(uint(i*2)), "")
+		localVars.SetDouble(uint8(i*2), v)
+		assert.Equal(t, v, localVars.GetDouble(uint8(i*2)), "")
 	}
 }
